requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
httpx==0.27.2
vonage==4.6.0